CHROMIUM: compat jail-control interface
We need a compat_ioctl function for 32-bit programs on 64-bit
kernels. In addition, since the struct size is different, we need
to define a compat struct and compat ioctl number for the
JAIL_CONTROL_ADD_DEVICE ioctl.

Note to future cherry-pickers: this can be squashed into the
previous device_jail commit to reduce the amount of patches we
need to carry forward.

BUG=chromium:644338
TEST=device_jail_utility and device_jail_fs on kevin

Change-Id: Id88644902526ca84cbbedb37cfd31c406f3d3a2d
Reviewed-on: https://chromium-review.googlesource.com/438768
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>