CHROMIUM: make device_jail depend on USB
device_jail explicitly depends on certain USB device-specific
functions and symbols and so we get undefined references if we
don't have USB enabled. Because device_jail operates on USB devices
for now, it doesn't make sense to compile it in unless we are
also compiling in USB anyway.

BUG=None
TEST=emerge-samus chromeos-kernel-3_14

Change-Id: I0ea19e6fe96f50e5966516b3dda1da75aaeecf91
Reviewed-on: https://chromium-review.googlesource.com/438806
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>