CHROMIUM: make device_jail depend on USB
device_jail explicitly depends on certain USB device-specific
functions and symbols and so we get undefined references if we
don't have USB enabled. Because device_jail operates on USB devices
for now, it doesn't make sense to compile it in unless we are
also compiling in USB anyway.

BUG=None
TEST=emerge-elm chromeos-kernel-3_18

Change-Id: I89d62fcc46e22746f6cc22897f970e0b3f73fee5
Reviewed-on: https://chromium-review.googlesource.com/438672
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>