CHROMIUM: drm/i915: Don't evict bound object in the shrinker
Before swapping, the shrinker gets called on all drivers to try
and reclaim some memory before resorting to disk access. For
i915, the shrinker will eventually try to move objects out of
the GTT to make them swappable. However when this happens too
often, this will have two negative effects:

- once we have exhausted all the purgeable objects, the
remaining objects will start bouncing in and out of the GTT
as the rendering drags them in and the shrinker pulls them out

- the i915 shrinker needs to wait on the CPU for the GPU to
complete pending rendering, which kills all asynchronicity

Those two things will slow down the system significantly as soon
as we start hitting the shrinker. So in this patch, we change
the shrinker to stop evicting in-use GTT objects. This avoids
both issues.

BUG=none
TEST=compiles and runs, seems to behave better under memory pressure

Change-Id: Ibfab2fa10c27dc25f33cf42e822bf07269a4bdfe
Reviewed-on: https://chromium-review.googlesource.com/439433
Commit-Ready: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>