CHROMIUM: iwl7000: ieee80211: fix SUITE B AKM definitions
The 11/12 are decimal, not hex - the original bug was
copied from wpa_supplicant but is already fixed there.

Fix this now - note that when sending upstream it has
to be changed to
	SUITE(0x000FAC, 11)
and
	SUITE(0x000FAC, 12)

respectively since I changed that there to make it
less likely to have the same mistake again.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 46c8a7d76edc470107cb99ac3283a5d1a64d8666