UPSTREAM: [media] cec: initiator should be the same as the destination for, poll
Poll messages that are used to allocate a logical address should
use the same initiator as the destination. Instead, it expected that
the initiator was 0xf which is not according to the standard.

This also had consequences for the message checks in cec_transmit_msg_fh
that incorrectly rejected poll messages with the same initiator and
destination.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>
(cherry picked from commit 42980da2eb7eb9695d8efc0c0ef145cbbb993b2c)
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>

Change-Id: I7d4a753e7336b6c776d5b8ec8495fae1d2709f9a
Reviewed-on: https://chromium-review.googlesource.com/1021307
Commit-Ready: Sean Paul <seanpaul@google.com>
Tested-by: Sean Paul <seanpaul@google.com>
Reviewed-by: Sean Paul <seanpaul@google.com>