CHROMIUM: arm64: dts: rockchip: Update dmc volts and idle states
This lowers the voltages for 800 and 928MHz for the dmc. It also
disables some of the idle states for 800+MHz since those were causing
ripple on the center logic rail. The ripple was causing issues even
with the increased voltage. This fixes the problem without the need to
increase the voltage as much at 928MHz.

BUG=chrome-os-partner:62855
TEST=on kevin, stressapptest -M 1500 -s 100000

Change-Id: Ib98fce44fe08644b37b1d0d19ec44459cff512c1
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/444253
Reviewed-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit 1aa67a5cd08c41c508f71489a32613a07f09e39f)
Reviewed-on: https://chromium-review.googlesource.com/450077
Tested-by: Douglas Anderson <dianders@chromium.org>