FROMLIST: btmrvl: wake system up when receives a wake irq
Currrently we are disabling this wake irq after receiving it. If this
happens before we finish suspend and the pm event check is disabled,
the system will continue suspending, and this irq would not work again.

We may need to abort system suspend to avoid that.

Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>
(am from https://patchwork.kernel.org/patch/9589453/)

BUG=b:35585605
TEST=Build and boot

Change-Id: I7a8c94cc1710a0a464dc78e5acc896e7ea58f374
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/445538
Reviewed-by: Brian Norris <briannorris@chromium.org>