UPSTREAM: drm/i915: Drop spinlocks around adding to the client request list
Adding to the tail of the client request list as the only other user is
in the throttle ioctl that iterates forwards over the list. It only
needs protection against deletion of a request as it reads it, it simply
won't see a new request added to the end of the list, or it would be too
early and rejected. We can further reduce the number of spinlocks
required when throttling by removing stale requests from the client_list
as we throttle.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Mika Kuoppala <mika.kuoppala@intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/20170302122525.19675-1-chris@chris-wilson.co.uk
(cherry picked from commit c8659efac544b7df1106022d78617a1ca1d9a38c)

BUG=b:131212476
TEST=Run cts tests to confirm vulkan 1.1 enabled on eve:
android.graphics.cts.VulkanFeaturesTest#testVulkan1_1Requirements
android.graphics.cts.VulkanFeaturesTest#testVulkanHardwareFeatures

Change-Id: I64fa50eb38cb93cff4fd739fa74eab4c7b3d8c66
Signed-off-by: Nataraj Deshpande <nataraj.deshpande@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1620029
Tested-by: Dmytro Chystiakov <dmytro.chystiakov@intel.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>