UPSTREAM: USB: serial: omninet: fix reference leaks at open
This driver needlessly took another reference to the tty on open, a
reference which was then never released on close. This lead to not just
a leak of the tty, but also a driver reference leak that prevented the
driver from being unloaded after a port had once been opened.

BUG=chromium:725018
TEST=Build and run

Change-Id: I9d01b5778bf6943dd457441fcb9ce35a7ea5b462
Fixes: 4a90f09b20f4 ("tty: usb-serial krefs")
Cc: stable <stable@vger.kernel.org>	# 2.6.28
Signed-off-by: Johan Hovold <johan@kernel.org>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit 30572418b445d85fc)
Reviewed-on: https://chromium-review.googlesource.com/514344
Reviewed-by: Dylan Reid <dgreid@chromium.org>