ANDROID: binder: use group leader instead of open thread
The binder allocator assumes that the thread that
called binder_open will never die for the lifetime of
that proc. That thread is normally the group_leader,
however it may not be. Use the group_leader instead
of current.

Bug: 35707103
Test: Created test case to open with temporary thread

BUG=None
TEST=None

Change-Id: Id693f74b3591f3524a8c6e9508e70f3e5a80c588
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: Martijn Coenen <maco@android.com>
(cherry picked from commit 8e5822fdabbb55608b62a100156e17612b590d39)
Signed-off-by: Prashant Malani <pmalani@google.com>
Reviewed-on: https://chromium-review.googlesource.com/958485
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>