ANDROID: android: binder: add padding to binder_fd_array_object.
binder_fd_array_object starts with a 4-byte header,
followed by a few fields that are 8 bytes when
ANDROID_BINDER_IPC_32BIT=N.

This can cause alignment issues in a 64-bit kernel
with a 32-bit userspace, as on x86_32 an 8-byte primitive
may be aligned to a 4-byte address. Pad with a __u32
to fix this.

BUG=None
TEST=None

Change-Id: I4374ed2cc3ccd3c6a1474cb7209b53ebfd91077b
Signed-off-by: Martijn Coenen <maco@android.com>
(cherry picked from commit a388cf91c7fd5279e923e167af59925826b055d5)
Signed-off-by: Prashant Malani <pmalani@google.com>
Reviewed-on: https://chromium-review.googlesource.com/958486
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>