CHROMIUM: iwl7000: mvm: SAR profiles should go from 1 to 4
The SAR profiles are officially defined as going from 1 to 4.  The
default is 1, which comes from WRDS.  And profiles 2 to 4 come from
the EWRD.

The driver currently uses 0 to 3, so we need to offset the numbers by
one when accessing our internal array.  Profile 0 (which is specified
as "disable SAR") is not supported.

Change-Id: I56fd2f6ba8865359cd593beb5eb522e86658cddb
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 6ed1bf6f15bbb6ef6c6229da843fd16814746ede
Reviewed-on: https://chromium-review.googlesource.com/459121
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>