BACKPORT: ASoC: Intel: Skylake: Add support for deferred DSP module bind
Module at the end of DSP pipeline that needs to be connected to a module
in another pipeline are represented as a PGA(leaf node) and in PGA event
handler these modules are bound/unbounded. Modules other than PGA leaf
can be connected directly or via switch to a module in another pipeline.
Example: reference path.

To support the deferred DSP module bind, following changes are done:
o When the path is enabled, the destination module that needs to be
bound may not be initialized. If the module is not initialized, add
these modules in a deferred bind list.
o When the destination module is initialized, check for these modules
in deferred bind list. If found, bind them.
o When the destination module is deleted, Unbind the modules.
o When the source module is deleted, remove the entry from the deferred
bind list.

Signed-off-by: Jeeja KP <jeeja.kp@intel.com>
Acked-by: Vinod Koul <vinod.koul@intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>

LINK:https://patchwork.kernel.org/patch/9677705/
BRANCH: for-next
(cherry picked from commit b8c722ddd548186db3f62e64f727af4db1cf9517)

BUG=b:37753089
TEST= build and boot. verify audio works.

Conflicts:
	sound/soc/intel/skylake/skl-topology.c
	sound/soc/intel/skylake/skl-topology.h

Change-Id: If515cbcdbca0cb2b5e1f58b2d2ee458843325dd2
Signed-off-by: Naveen M <naveen.m@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/540037
Commit-Ready: Sathya Prakash <sathya.prakash.m.r@intel.corp-partner.google.com>
Tested-by: Sathya Prakash <sathya.prakash.m.r@intel.corp-partner.google.com>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>