BACKPORT: HID: sony: Make DS4 bt poll interval adjustable
By default when using bluetooth the DS4 reports data at about 1kHz,
which is quite fast especially on weak devices. We now make the
device use the USB poll interval, which is a fixed 4ms. In addition
we make the value adjustable through sysfs.

The error handling in sony_input_configured is a little tricky. It
is not easy to add other goto's as not all codepaths have logic
for adding this attribute. Luckily we are setting the value for the
attribute to a default value, so we can use that to detect if we need
to remove the file.

Signed-off-by: Roderick Colenbrander <roderick.colenbrander@sony.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit 77b499e739ed5561e5026fa7140ae53f6c4d1d8e)

Conflicts:
   drivers/hid/hid-sony.c

BUG=b:139939892, chromium:972297
TEST=DualShock v4 via Bluetooth works on https://html5gamepad.com/

Change-Id: Ia4c68ee71699eb93c49abeb0d4797097abce81ed
Signed-off-by: Tetsui Ohkubo <tetsui@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1800281
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>