FROMLIST: mac80211: Fix clang warning about constant operand in logical operation
When clang detects a non-boolean constant in a logical operation it
generates a 'constant-logical-operand' warning. In
ieee80211_try_rate_control_ops_get() the result of strlen(<const str>)
is used in a logical operation, clang resolves the expression to an
(integer) constant at compile time when clang's builtin strlen function
is used.

Change the condition to check for strlen() > 0 to make the constant
operand boolean and thus avoid the warning.

Link: https://patchwork.kernel.org/patch/9668641/
(applied to git.kernel.org jberg/mac80211-next master)

BUG=chromium:702741
TEST=build with clang for gru

Change-Id: I9f9bd2fad8c792aaec2ca2d1bdbadb34bae0ec1d
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/477155
Reviewed-by: Guenter Roeck <groeck@chromium.org>